% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickplots.R
\name{default_income_groups}
\alias{default_income_groups}
\title{Default income groups, just to avoid repeating so many lines in all functions}
\usage{
default_income_groups()
}
\description{
Default income groups, just to avoid repeating so many lines in all functions
}
