% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickplots.R
\name{default_regions}
\alias{default_regions}
\title{Default regions, just to avoid repeating so many lines in all functions}
\usage{
default_regions()
}
\description{
Default regions, just to avoid repeating so many lines in all functions
}
