% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickplots.R
\name{get_char_code}
\alias{get_char_code}
\title{Return special character codes}
\usage{
get_char_code(col_name)
}
\arguments{
\item{col_name}{name of the column indicator; ind_1, ind2, ..., ind_n}
}
\description{
Auxiliary function to return superscript numbers
}
\keyword{internal}
