% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickplots.R
\name{get_formatter}
\alias{get_formatter}
\title{Create custom formatter for indicators, based on the data and label}
\usage{
get_formatter(plot_data)
}
\arguments{
\item{plot_data}{indicator data, including label}
}
\value{
a formatter function from the scales package
}
\description{
Create custom formatter for indicators, based on the data and label
}
\keyword{internal}
