% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickplots.R
\name{interpolate_wdi}
\alias{interpolate_wdi}
\title{Interpolate missing values in indicator variables}
\usage{
interpolate_wdi(wdi_data)
}
\arguments{
\item{wdi_data}{dataset to interpolate, as returned by download_wdi}
}
\value{
a data.frame with interpolated values and additional column
        indicating which indicators, for which year and country were
        interpolated
}
\description{
most probably, there will be missing values in some countries for some years
so here's a controversial decision to make the animations look good. Let's
fill missing values for each country, by interpolating values in the gaps and
filling with the first or last value available
}
