% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickplots.R
\name{latest_wdi_ind}
\alias{latest_wdi_ind}
\title{Get latest data for a single indicator per country}
\usage{
latest_wdi_ind(
  indicator = "NY.GDP.PCAP.KD",
  highlight_countries = c(""),
  start = lubridate::year(Sys.Date()) - 10,
  end = lubridate::year(Sys.Date()),
  country = "all"
)
}
\arguments{
\item{indicator}{character of length 1 with the indicator code}

\item{highlight_countries}{character vector with country names to highlight}

\item{start}{first year to download data}

\item{end}{last year to download data}

\item{country}{subset of countries to download data. Default to "all"}
}
\value{
data.frame with columns country, year, plot_ind, region, income,
        highlight. highlight = NA, except for highlight countries in which
        case takes the value of plot_ind
}
\description{
Get latest data for a single indicator per country
}
\examples{
\dontrun{
latest_wdi_ind("NY.GDP.PCAP.KD", c("Colombia", "Germany"), 2019, 2019)
}

}
