% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickplots.R
\name{plot_dist_wdi_ggpdef}
\alias{plot_dist_wdi_ggpdef}
\title{Plot a WDI indicator}
\usage{
plot_dist_wdi_ggpdef(wdi_data, ind, facets, country, highlight, year, p = 1)
}
\arguments{
\item{wdi_data}{data.frame as returned by latest_wdi}

\item{ind}{variable in wdi_data to plot}

\item{facets}{variable in wdi_data to use as facets (either region or income)}

\item{country}{variable in wdi_data to use as country names}

\item{highlight}{variable in wdi_data with the value of countries to highglight}

\item{year}{variable in wdi_data with the value of the year for each point}

\item{p}{Transformation exponent, <U+03BB>, as in scales::modulus_trans}
}
\value{
a ggplot2 object
}
\description{
Plot a WDI indicator
}
