% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickplots.R
\name{plot_dist_wdi_ind}
\alias{plot_dist_wdi_ind}
\title{Plot the distribution (density) of a single indicator using WDI data,
downloading the data as necessary for the period indicated by start and end,
and keeping only the latest data point available for each country}
\usage{
plot_dist_wdi_ind(
  indicator = "NY.GDP.PCAP.KD",
  highlight_countries = c("Colombia", "Germany"),
  groups = region,
  start = lubridate::year(Sys.Date()) - 10,
  end = lubridate::year(Sys.Date()),
  country = "all",
  p = 0
)
}
\arguments{
\item{indicator}{character of length 1 with the indicator code}

\item{highlight_countries}{character vector with country names to highlight}

\item{groups}{whether to show facets per region or income}

\item{start}{first year to download data}

\item{end}{last year to download data}

\item{country}{subset of countries to download data. Default to "all"}

\item{p}{Transformation exponent, λ, as in scales::modulus_trans}
}
\value{
a ggplot2 object
}
\description{
Plot the distribution (density) of a single indicator using WDI data,
downloading the data as necessary for the period indicated by start and end,
and keeping only the latest data point available for each country
}
\examples{
\dontrun{
plot_dist_wdi_ind()
}
}
