% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickplots.R
\name{plot_race_wdi}
\alias{plot_race_wdi}
\title{Race bar plot}
\usage{
plot_race_wdi(
  indicator = "SI.POV.GINI",
  highlight_countries = c("Colombia", "Germany"),
  start = lubridate::year(Sys.Date()) - 15,
  end = lubridate::year(Sys.Date() - months(18)),
  country = "all",
  regions = default_regions(),
  income_groups = default_income_groups(),
  p = 1
)
}
\arguments{
\item{indicator}{character of length 1 with the indicator code}

\item{highlight_countries}{character vector with country names to highlight}

\item{start}{first year to download data}

\item{end}{last year to download data}

\item{country}{subset of countries to download data from. Default to "all".
This argument is passed as is to WDI::WDI and hence, requires
'ISO-2 character codes, e.g. "BR", "US", "CA"'.}

\item{regions}{character vector to filter the data only to specific regions}

\item{income_groups}{character vector to filter the data only to specific
income groups}

\item{p}{transformation exponent, as in scales::modulus_trans}
}
\value{
gganimate
}
\description{
Race bar plot
}
\examples{
\dontrun{
plot_race_wdi()
}
}
