% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickplots.R
\name{plot_race_wdi_ind}
\alias{plot_race_wdi_ind}
\title{Race bar plot}
\usage{
plot_race_wdi_ind(
  indicator = "SI.POV.GINI",
  highlight_countries = c("Colombia", "Germany"),
  start = lubridate::year(Sys.Date()) - 10,
  end = lubridate::year(Sys.Date()),
  country = "all",
  regions = wdiquickplots::regions,
  income_groups = wdiquickplots::income_groups
)
}
\arguments{
\item{indicator}{character of length 1 with the indicator code}

\item{highlight_countries}{character vector with country names to highlight}

\item{start}{first year to download data}

\item{end}{last year to download data}

\item{country}{subset of countries to download data from. Default to "all".
This argument is passed as is to WDI::WDI and hence, requires
'ISO-2 character codes, e.g. "BR", "US", "CA"'.}

\item{regions}{character vector to filter the data only to specific regions}

\item{income_groups}{character vector to filter the data only to specific
sincome groups}
}
\value{
gganimate
}
\description{
Race bar plot
}
