% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickplots.R
\name{plot_spaghetti_wdi_ind}
\alias{plot_spaghetti_wdi_ind}
\title{Spaghetti plot that no-one really want to see (nor should want to see, ever)}
\usage{
plot_spaghetti_wdi_ind(
  indicator = "SI.POV.GINI",
  highlight_countries = c("Colombia", "Germany"),
  start = lubridate::year(Sys.Date()) - 10,
  end = lubridate::year(Sys.Date()),
  country = "all"
)
}
\arguments{
\item{indicator}{character of length 1 with the indicator code}

\item{highlight_countries}{character vector with country names to highlight}

\item{start}{first year to download data}

\item{end}{last year to download data}

\item{country}{subset of countries to download data. Default to "all"}
}
\value{
a dygraphs plot
}
\description{
Spaghetti plot that no-one really want to see (nor should want to see, ever)
}
