% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickplots.R
\name{plot_time_facets_wdi}
\alias{plot_time_facets_wdi}
\title{Line plot with facets showing how the indicator has changed over time
in the highlight_countries, within regions or income groups (facets)}
\usage{
plot_time_facets_wdi(
  indicator = "SI.POV.GINI",
  highlight_countries = c("Colombia", "Germany"),
  facets = region,
  start = lubridate::year(Sys.Date()) - 10,
  end = lubridate::year(Sys.Date() - months(18)),
  country = "all",
  regions = default_regions(),
  income_groups = default_income_groups(),
  p = 1
)
}
\arguments{
\item{indicator}{character of length 1 with the indicator code}

\item{highlight_countries}{character vector with country names to highlight}

\item{facets}{variable to use for facets. Either region or income}

\item{start}{first year to download data}

\item{end}{last year to download data}

\item{country}{subset of countries to download data from. Default to "all".
This argument is passed as is to WDI::WDI and hence, requires
'ISO-2 character codes, e.g. "BR", "US", "CA"'.}

\item{regions}{character vector to filter the data only to specific regions}

\item{income_groups}{character vector to filter the data only to specific
income groups}

\item{p}{Transformation exponent, <U+03BB>, as in scales::modulus_trans}
}
\value{
ggplot object
}
\description{
We use the super cool gghighlight::gghighlight package to disentangle the
spaghetti plot and highlight selected countries in each facet
}
\examples{
\dontrun{
plot_time_facets_wdi()
}
}
