% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickplots.R
\name{plot_time_wdi_ind}
\alias{plot_time_wdi_ind}
\title{Line plot (plotly-powered) the indicator over time (year) for the
highlight_countries only}
\usage{
plot_time_wdi_ind(
  indicator = "SI.POV.GINI",
  highlight_countries = c("Colombia", "Germany"),
  start = lubridate::year(Sys.Date()) - 10,
  end = lubridate::year(Sys.Date()),
  country = "all"
)
}
\arguments{
\item{indicator}{character of length 1 with the indicator code}

\item{highlight_countries}{character vector with country names to highlight}

\item{start}{first year to download data}

\item{end}{last year to download data}

\item{country}{subset of countries to download data. Default to "all"}
}
\value{
plotly object
}
\description{
Line plot (plotly-powered) the indicator over time (year) for the
highlight_countries only
}
