% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickplots.R
\name{plot_wdi_dist_col}
\alias{plot_wdi_dist_col}
\title{Plot a WDI indicator}
\usage{
plot_wdi_dist_col(wdi_data, ind, groups, country, highlight, p = 0)
}
\arguments{
\item{wdi_data}{data.frame as returned by latest_wdi_ind}

\item{ind}{variable in wdi_data to plot}

\item{groups}{variable in wdi_data to use as facets (either region or income)}

\item{country}{variable in wdi_data to use as country names}

\item{highlight}{variable in wdi_data with the value of countries to highglight}

\item{p}{Transformation exponent, λ, as in scales::modulus_trans}
}
\value{
a ggplot2 object
}
\description{
Plot a WDI indicator
}
\examples{
\dontrun{
plot_wdi_dist_col()
}
}
