% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdiquickplots-package.R
\docType{package}
\name{wdiquickplots-package}
\alias{wdiquickplots}
\alias{wdiquickplots-package}
\title{wdiquickplots: Quick Plots for WDI data and other World Bank-hosted databases}
\description{
Plot indicators from WDI data with one line of code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/edalfon/wdiquickplots}
  \item Report bugs at \url{https://github.com/edalfon/wdiquickplots/issues}
}

}
\keyword{internal}
