% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickplots.R
\name{download_wdi_ind}
\alias{download_wdi_ind}
\title{Download WDI data and prepare the dataset for the plots}
\usage{
download_wdi_ind(
  indicator = "NY.GDP.PCAP.KD",
  highlight_countries = c(""),
  start = lubridate::year(Sys.Date()) - 10,
  end = lubridate::year(Sys.Date()),
  country = "all",
  regions = default_regions(),
  income_groups = default_income_groups(),
  interpolate = FALSE
)
}
\arguments{
\item{indicator}{character of length 1 with the indicator code}

\item{highlight_countries}{character vector with country names to highlight}

\item{start}{first year to download data}

\item{end}{last year to download data}

\item{country}{subset of countries to download data from. Default to "all".
This argument is passed as is to WDI::WDI and hence, requires
'ISO-2 character codes, e.g. "BR", "US", "CA"'.}

\item{regions}{character vector to filter the data only to specific regions}

\item{income_groups}{character vector to filter the data only to specific
income groups}

\item{interpolate}{boolean to indicate whether to interpolate missing values}
}
\value{
data.frame with columns country, year, ind_1, region, income,
        highlight_ind_1. highlight_ind_1 = NA, except for highlight
        countries in which case, it takes the value of ind_1
}
\description{
Download data using a memoised version of WDI::WDI() and wrangle a bit
with the data to prepare a data.frame for the plots#'
}
\examples{
\dontrun{
download_wdi_ind("NY.GDP.PCAP.KD", c("Colombia", "Germany"), 2019, 2019)
}

}
