% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickplots.R
\name{plot_bar_wdi_ind}
\alias{plot_bar_wdi_ind}
\title{Plot a WDI indicator as an interactive bar-plot (powered by plotly)}
\usage{
plot_bar_wdi_ind(
  indicator = "NY.GDP.PCAP.KD",
  highlight_countries = c("Colombia", "Germany"),
  start = lubridate::year(Sys.Date()) - 10,
  end = lubridate::year(Sys.Date()),
  country = "all",
  regions = wdiquickplots:::regions,
  income_groups = wdiquickplots:::income_groups,
  base_color = "skyblue",
  highlight_color = "red"
)
}
\arguments{
\item{indicator}{character of length 1 with the indicator code}

\item{highlight_countries}{character vector with country names to highlight}

\item{start}{first year to download data}

\item{end}{last year to download data}

\item{country}{subset of countries to download data from. Default to "all".
This argument is passed as is to WDI::WDI and hence, requires
'ISO-2 character codes, e.g. "BR", "US", "CA"'.}

\item{regions}{character vector to filter the data only to specific regions}

\item{income_groups}{character vector to filter the data only to specific
sincome groups}

\item{base_color}{fill color for the bars for most countries}

\item{highlight_color}{fill color for the bars for highlighted countries}
}
\value{
a plotly object
}
\description{
Plot a WDI indicator as an interactive bar-plot (powered by plotly)
}
\examples{
\dontrun{
plot_bar_wdi_ind()
}
}
