% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickplots.R
\name{plot_dist_wdi_ind}
\alias{plot_dist_wdi_ind}
\title{Plot the distribution (density) of a single indicator using WDI data,
downloading the data as necessary for the period indicated by start and end,
and keeping only the latest data point available for each country}
\usage{
plot_dist_wdi_ind(
  indicator = "NY.GDP.PCAP.KD",
  highlight_countries = c("Colombia", "Germany"),
  facets = region,
  start = lubridate::year(Sys.Date()) - 10,
  end = lubridate::year(Sys.Date()),
  country = "all",
  regions = default_regions(),
  income_groups = default_income_groups(),
  p = 1
)
}
\arguments{
\item{indicator}{character of length 1 with the indicator code}

\item{highlight_countries}{character vector with country names to highlight}

\item{facets}{whether to show facets per region or income}

\item{start}{first year to download data}

\item{end}{last year to download data}

\item{country}{subset of countries to download data from. Default to "all".
This argument is passed as is to WDI::WDI and hence, requires
'ISO-2 character codes, e.g. "BR", "US", "CA"'.}

\item{regions}{character vector to filter the data only to specific regions}

\item{income_groups}{character vector to filter the data only to specific
income groups}

\item{p}{Transformation exponent, <U+03BB>, as in scales::modulus_trans}
}
\value{
a ggplot2 object
}
\description{
Plot the distribution (density) of a single indicator using WDI data,
downloading the data as necessary for the period indicated by start and end,
and keeping only the latest data point available for each country
}
\examples{
\dontrun{
plot_dist_wdi_ind()
}
}
