% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickplots.R
\name{plot_time_facets_wdi_ind}
\alias{plot_time_facets_wdi_ind}
\title{Line plot with facets showing how the indicator has changed over time
in the highlight_countries, within regions or income groups (facets)}
\usage{
plot_time_facets_wdi_ind(
  indicator = "SI.POV.GINI",
  highlight_countries = c("Colombia", "Germany"),
  facets = region,
  start = lubridate::year(Sys.Date()) - 10,
  end = lubridate::year(Sys.Date()),
  country = "all",
  p = 0
)
}
\arguments{
\item{indicator}{character of length 1 with the indicator code}

\item{highlight_countries}{character vector with country names to highlight}

\item{facets}{variable to use for facets. Either region or income}

\item{start}{first year to download data}

\item{end}{last year to download data}

\item{country}{subset of countries to download data. Default to "all"}

\item{p}{Transformation exponent, <U+03BB>, as in scales::modulus_trans}
}
\value{
ggplot object
}
\description{
We use the super cool gghighlight::gghighlight package to disentangle the
spaghetti plot and highlight selected countries in each facet
}
