% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdiquickplots-package.R
\docType{package}
\name{wdiquickplots-package}
\alias{wdiquickplots}
\alias{wdiquickplots-package}
\title{wdiquickplots: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
